+++ pkg/microservice/aslan/core/common/util/istio_grayscale.go
@@ -5,7 +5,7 @@ Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
 
-   http://www.apache.org/licenses/LICENSE-2.0
+  http://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,17 +14,16 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package service
+package util
 
 import (
 	"context"
+	zadigutil "github.com/koderover/zadig/pkg/util"
 
 	commonmodels "github.com/koderover/zadig/pkg/microservice/aslan/core/common/repository/models"
 	commonrepo "github.com/koderover/zadig/pkg/microservice/aslan/core/common/repository/mongodb"
-	zadigutil "github.com/koderover/zadig/pkg/util"
 )
-
-func ensureIstioGrayConfig(ctx context.Context, baseEnv *commonmodels.Product) error {
+func EnsureIstioGrayConfig(ctx context.Context, baseEnv *commonmodels.Product) error {
 	if baseEnv.IstioGrayscale.Enable && baseEnv.IstioGrayscale.IsBase {
 		return nil
 	}
@@ -37,7 +36,7 @@ func ensureIstioGrayConfig(ctx context.Context, baseEnv *commonmodels.Product) e
 	return commonrepo.NewProductColl().Update(baseEnv)
 }
 
-func fetchGrayEnvs(ctx context.Context, productName, clusterID, baseEnvName string) ([]*commonmodels.Product, error) {
+func FetchGrayEnvs(ctx context.Context, productName, clusterID, baseEnvName string) ([]*commonmodels.Product, error) {
 	return commonrepo.NewProductColl().List(&commonrepo.ProductListOptions{
 		Name:                  productName,
 		ClusterID:             clusterID,
